q2w bsp format support. automatic lightmap texture scaling to retain more performance on large maps. r_clutter preliminary implementation should probably fix up the shader still. CSQC_Parse_Damage implemented. finally implement q2 inventory. fix mixer overflow crash. glsl can now use s_diffuse etc to force inclusion of a diffuse sampler/texture, meaning shaders don't need to include them. fix issue with writeip
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4841 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/vid_headless.c
@@ -187,13 +187,13 @@ static qboolean Headless_BE_LightCullModel	(vec3_t org, struct model_s *model)
 {
 	return false;
 }
-static void Headless_BE_VBO_Begin	(vbobctx_t *ctx, unsigned int maxsize)
+static void Headless_BE_VBO_Begin	(vbobctx_t *ctx, size_t maxsize)
 {
 }
-static void Headless_BE_VBO_Data	(vbobctx_t *ctx, void *data, unsigned int size, vboarray_t *varray)
+static void Headless_BE_VBO_Data	(vbobctx_t *ctx, void *data, size_t size, vboarray_t *varray)
 {
 }
-static void Headless_BE_VBO_Finish	(vbobctx_t *ctx, void *edata, unsigned int esize, vboarray_t *earray)
+static void Headless_BE_VBO_Finish	(vbobctx_t *ctx, void *edata, size_t esize, vboarray_t *earray)
 {
 }
 static void Headless_BE_VBO_Destroy	(vboarray_t *vearray)
